+++ src/main/java/frc/robot/subsystems/LEDSubsystem.java
@@ -1,105 +1,149 @@
 package frc.robot.subsystems;
-
-
-import java.util.Optional;
-
-import edu.wpi.first.wpilibj.AddressableLED;
-import edu.wpi.first.wpilibj.AddressableLEDBuffer;
-import edu.wpi.first.wpilibj.DriverStation;
-import edu.wpi.first.wpilibj.DriverStation.Alliance;
-import edu.wpi.first.wpilibj.util.Color;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import frc.robot.Constants.LED;
+import com.ctre.phoenix.led.CANdle;
+import com.ctre.phoenix.led.RainbowAnimation;
+import com.ctre.phoenix.led.LarsonAnimation;
+import com.ctre.phoenix.led.StrobeAnimation;
+import com.ctre.phoenix.led.CANdle.LEDStripType;
+import frc.robot.subsystems.NoteSubSystem.Target;;
 
-public class LEDSubsystem extends SubsystemBase {
-  private final AddressableLED m_led = new AddressableLED(LED.PWMPORT);
-  private final AddressableLEDBuffer m_ledBuffer = new AddressableLEDBuffer(LED.BUFFERSIZE);
 
-  private boolean shooter = false;
-  private boolean redBall = false;
-  private boolean blueBall = false;
+public class LEDSubSystem extends SubsystemBase {
+    CANdle candle1 = new CANdle(60);
+    //create a rainbow anim.
+    RainbowAnimation rainbowAnimation = new RainbowAnimation(1, 1, 40);
 
-  public LEDSubsystem() {
-    m_led.setLength(m_ledBuffer.getLength());
-    m_led.setData(m_ledBuffer);
-    m_led.start();
-  }
+    // LarsonAnimation manualModeSegment1 = new LarsonAnimation(0,0,0,0,1,10,0,3,8);
+    //int r, int g, int b, int w, double speed, int numLed, BounceMode mode, int size, int ledOffset)
 
-  @Override
-  public void periodic() {
-    if (DriverStation.isEnabled()) {
-      Optional<Alliance> ally = DriverStation.getAlliance();
-      if (ally.equals(Alliance.Red)) {
-
-        redBall = true;
-      }
-      else if (ally.equals(Alliance.Blue)) { // should only have pipelines 0 & 1
-
-        blueBall = true;
-      }
-      setLEDs();
+   
+    public LEDSubSystem() {
+        candle1.configLEDType(LEDStripType.RGB);
+        candle1.clearAnimation(0);
+        candle1.configBrightnessScalar(1);
     }
-  }
 
-  @Override
-  public void simulationPeriodic() {}
+    @Override
+    public void periodic() {
+        // This method will be called once per scheduler run
 
-  public void setShooter(boolean ready){
-    shooter=ready;
-  }
+    }
 
-  private void setLEDs() {
-    setBallLEDs();
-    setShooterLEDs();
-    m_led.setData(m_ledBuffer);
-  }
+    @Override
+    public void simulationPeriodic() {
+        // This method will be called once per scheduler run when in simulation
 
-  private void setBallLEDs() {
-    if (redBall && blueBall) {
-      setFrontHalf();
     }
-    else if (redBall || blueBall) {
-      if (redBall) {
-        setFrontAll(Color.kRed);
-      }
-      if (blueBall) {
-        setFrontAll(Color.kBlue);
-      }
-    }
-    else {
-      setFrontAll(Color.kBlack); // Off
-    }
-  }
 
-  private void setShooterLEDs() {
-    if (shooter) {
-      setBackAll(Color.kGreen);
+    public void setEmpty(){
+        //all solid orange
+        candle1.setLEDs(255, 24, 0, 0, 0, 48);
     }
-    else {
-      setBackAll(Color.kBlack); // Off
-    }
-  }
 
-  private void setFrontAll(Color color) {
-    for (var i = 0; i < m_ledBuffer.getLength() / 2; i++) {
-      m_ledBuffer.setLED(i, color);
+    public void setIntaking(){
+        //orange rolling
     }
-  }
 
-  public void setFrontHalf() {
-    for (int i = 0; i < m_ledBuffer.getLength() / 2; i++) {
-      if (i < m_ledBuffer.getLength() / 2) {
-        m_ledBuffer.setLED(i, Color.kBlue);
-      }
-      else {
-        m_ledBuffer.setLED(i, Color.kRed);
-      }
+    public void setHaveNote(){
+        //solid blue
+        candle1.setLEDs(0, 0, 255, 0, 0, 48);
     }
-  }
 
-  public void setBackAll(Color color) {
-    for (var i = m_ledBuffer.getLength() / 2; i < m_ledBuffer.getLength(); i++) {
-      m_ledBuffer.setLED(i, color);
+    public void setPrepareToShoot(Target mytarget, boolean isVisionMode){
+        //blue rolling
+        // for manual mode, this is when: waiting for angle to be at its setpoint
+        //                                or for shooter to be up to speed
+        // for vision mode, this also when: target not found yet,
+        //                                  or outside distance we can shoot 
+        //                                  or heading is still not correct
+
+        //use the top LED to indicate manual mode target
+        //red = speaker at the subwoofer
+        //red blinking = speaker at podium
+        //yellow = amp
+        //orange = trap
+        //white = other
+
+        //top 3 LED lite means vision mode active
+        // red = speaker,  
+        // yellow = amp
+        // orange = trap
+        // all 3 top blink if target not found
+        //                 target found, top is solid
+        //                 within distance, 2nd is solid
+        //                 heading locked on, 3rd solid  (and remaining solid green)
+
+
+
     }
-  }
+
+    public void setReadyToShoot(Target mytarget, boolean isVisionMode){
+        //green solid
+        //this is at the setpoint angle and speed when shooting manually
+        //    or in the case of vision, also within distance and the heading is correct
+
+        //use the top LED to indicate manual mode target
+        //red = speaker at the subwoofer
+        //green = speaker at podium
+        //yellow = amp
+        //orange = trap
+        //white = other
+
+        //top 3 LED lite means vision mode
+        // red = speaker
+        // yellow = amp
+        // orange = trap
+
+        
+
+
+    }
+
+
+
+
+    public void candlePurple(){
+        //set brightness
+        candle1.configBrightnessScalar(1);
+        //set color
+        candle1.setLEDs(80, 45, 127);
+
+    }
+
+    public void candleGold(){
+        //set brightness
+        candle1.configBrightnessScalar(1);
+        //set color
+        candle1.setLEDs(255, 200, 46);
+    }
+    
+    public void candleOrange(){
+        //set brightness
+        candle1.configBrightnessScalar(1);
+        //set color
+        candle1.setLEDs(255, 24, 0, 0, 48, 10);
+        candle1.setLEDs(255, 24, 0, 0, 58, 10);
+        candle1.setLEDs(255, 24, 0, 0, 68, 10);
+        candle1.setLEDs(255, 24, 0, 0, 78, 10);
+        //237,125,49 - robbie
+        //230,126,34
+    }
+
+    public void candleBlue(){
+        //set brightness
+        candle1.configBrightnessScalar(1);
+        //set color
+        candle1.setLEDs(0, 0, 255, 0, 0, 8);
+        candle1.setLEDs(0, 0, 255, 0, 8, 40);
+        // 39,59,140 - robbie
+        //0,0,255
+    }
+
+    public void candleRainbow(){
+        //set brightness
+        candle1.configBrightnessScalar(1);
+        //set color
+        candle1.animate(rainbowAnimation);
+    }
 }
+
